+++ src/Futhark/CodeGen/Backends/HIP.hs
@@ -84,13 +84,6 @@ cliOptions =
              optionArgument = RequiredArgument "OPT",
              optionDescription = "Add an additional build option to the string passed to NVRTC.",
              optionAction = [C.cstm|futhark_context_config_add_build_option(cfg, optarg);|]
-           },
-         Option
-           { optionLongName = "profile",
-             optionShortName = Just 'P',
-             optionArgument = NoArgument,
-             optionDescription = "Gather profiling data while executing and print out a summary at the end.",
-             optionAction = [C.cstm|futhark_context_config_set_profiling(cfg, 1);|]
            }
        ]