+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Fusion across slicing is now possible in some cases.
 
+* New tool: `futhark profile`.
+
 ### Removed
 
 ### Changed